Extract duplicated code into class LiteralByMatchability
