+++ Indexing/LiteralIndex.cpp
@@ -23,6 +23,7 @@
 
 #include "Kernel/Clause.hpp"
 #include "Kernel/LiteralComparators.hpp"
+#include "Kernel/LiteralByMatchability.hpp"
 #include "Kernel/Matcher.hpp"
 #include "Kernel/MLVariant.hpp"
 #include "Kernel/Ordering.hpp"
@@ -111,85 +112,6 @@ void SimplifyingLiteralIndex::handleClause(Clause* c, bool adding)
   }
 }
 
-/**
- * A literal and its rating (for use in the subsumption index).
- * Ordered by ratings, breaking ties with the literal's id.
- *
- * The metric used to select the best literal is defined as follows:
- *
- *    val == #symbols - #distinct-variables
- *        == #non-variable-symbols + #variable-duplicates
- *
- * This value is the number of symbols that induce constraints for matching.
- * Note that variables only induce constraints for instantiation on their
- * repeated occurrences.  We want to maximize this value to have the most
- * restricting literal, so we get as little candidate clauses as possible
- * (because the candidates then have to be passed to the MLMatcher which is
- * expensive).
- */
-class SubsRatedLiteral
-{
-  private:
-    Literal* m_lit;
-    unsigned m_val;
-
-  public:
-    SubsRatedLiteral(Literal* lit)
-      : m_lit(lit), m_val(computeRating(lit))
-    { }
-
-    static unsigned computeRating(Literal* lit)
-    {
-      return lit->weight() - lit->getDistinctVars();
-    }
-
-    Literal* lit() const { return m_lit; }
-
-    bool operator<(SubsRatedLiteral const& other) const
-    {
-      return m_val < other.m_val || (m_val == other.m_val && m_lit->getId() < other.m_lit->getId());
-    }
-    bool operator>(SubsRatedLiteral const& other) const { return other.operator<(*this); }
-    bool operator<=(SubsRatedLiteral const& other) const { return !operator>(other); }
-    bool operator>=(SubsRatedLiteral const& other) const { return !operator<(other); }
-    bool operator==(SubsRatedLiteral const& other) const { return m_lit == other.m_lit; }
-    bool operator!=(SubsRatedLiteral const& other) const { return !operator==(other); }
-
-    static SubsRatedLiteral find_best_in(Clause* c)
-    {
-      ASS_GE(c->length(), 1);
-      SubsRatedLiteral best{(*c)[0]};
-      for (unsigned i = 1; i < c->length(); ++i) {
-        SubsRatedLiteral curr{(*c)[i]};
-        if (curr > best) {
-          best = curr;
-        }
-      }
-      return best;
-    }
-
-    static std::pair<SubsRatedLiteral,SubsRatedLiteral> find_best2_in(Clause* c)
-    {
-      ASS_GE(c->length(), 2);
-      SubsRatedLiteral best{(*c)[0]};
-      SubsRatedLiteral secondBest{(*c)[1]};
-      if (secondBest > best) {
-        std::swap(best, secondBest);
-      }
-      for (unsigned i = 2; i < c->length(); ++i) {
-        SubsRatedLiteral curr{(*c)[i]};
-        if (curr > best) {
-          secondBest = best;
-          best = curr;
-        } else if (curr > secondBest) {
-          secondBest = curr;
-        }
-      }
-      ASS(best.lit() != secondBest.lit());
-      ASS(best > secondBest);
-      return {best, secondBest};
-    }
-};
 
 void FwSubsSimplifyingLiteralIndex::handleClause(Clause* c, bool adding)
 {
@@ -201,7 +123,7 @@ void FwSubsSimplifyingLiteralIndex::handleClause(Clause* c, bool adding)
 
   TimeCounter tc(TC_FORWARD_SUBSUMPTION_INDEX_MAINTENANCE);
 
-  Literal* best = SubsRatedLiteral::find_best_in(c).lit();
+  Literal* best = LiteralByMatchability::find_least_matchable_in(c).lit();
   handleLiteral(best, c, adding);
 }
 
@@ -228,7 +150,7 @@ void FSDLiteralIndex::handleClause(Clause* c, bool adding)
     return;
   }
 
-  auto res = SubsRatedLiteral::find_best2_in(c);
+  auto res = LiteralByMatchability::find_two_least_matchable_in(c);
   Literal* best = res.first.lit();
   Literal* secondBest = res.second.lit();
   if (!best->isEquality() || !best->isPositive()) {
